--- a/pkg/services/ngalert/notifier/alertmanager_test.go
+++ b/pkg/services/ngalert/notifier/alertmanager_test.go
@@ -391,9 +391,9 @@ func TestSilenceCleanup(t *testing.T) {
 		// Active now
 		makeSilence("", "tests", dt(now.Add(-5*time.Hour)), dt(now.Add(6*time.Hour)), matchers),
 		// Expiring soon.
-		makeSilence("", "tests", dt(now.Add(-5*time.Hour)), dt(now.Add(5*time.Second)), matchers),
-		// Expiring *very* soon
 		makeSilence("", "tests", dt(now.Add(-5*time.Hour)), dt(now.Add(2*time.Second)), matchers),
+		// Expiring *very* soon
+		makeSilence("", "tests", dt(now.Add(-5*time.Hour)), dt(now.Add(1*time.Second)), matchers),
 	}
 
 	for _, s := range silences {
@@ -407,12 +407,12 @@ func TestSilenceCleanup(t *testing.T) {
 		found, err := am.ListSilences(nil)
 		require.NoError(err)
 		return len(found) == 3
-	}, 3*time.Second, 150*time.Millisecond)
+	}, 3*time.Second, 100*time.Millisecond)
 
 	// Wait again for another silence to expire.
 	require.Eventually(func() bool {
 		found, err := am.ListSilences(nil)
 		require.NoError(err)
 		return len(found) == 2
-	}, 6*time.Second, 150*time.Millisecond)
+	}, 6*time.Second, 100*time.Millisecond)
 }